refactor(client): move pos slice into features/pos
