+++ client/src/features/pos/components/StartupPrompt.tsx
@@ -8,13 +8,13 @@ import {
   DialogHeader,
   DialogTitle,
   DialogDescription,
-} from '../shared/ui/dialog';
-import { Button } from '../shared/ui/button';
-import { Input } from '../shared/ui/input';
-import { useAuthStore } from '../features/auth/store/authStore';
-import { useTranslation } from '../shared/i18n/index';
-import { resource } from '../shared/lib/resource';
-import { useTransport } from '../shared/lib/transport/index';
+} from '../../../shared/ui/dialog';
+import { Button } from '../../../shared/ui/button';
+import { Input } from '../../../shared/ui/input';
+import { useAuthStore } from '../../auth/store/authStore';
+import { useTranslation } from '../../../shared/i18n/index';
+import { resource } from '../../../shared/lib/resource';
+import { useTransport } from '../../../shared/lib/transport/index';
 import type { RegisterSession, Shift } from '@/types';
 
 const SESSION_KEY = 'moon-startup-dismissed';
