+++ client/src/features/pos/pages/BarcodeTools.tsx
@@ -2,19 +2,19 @@ import { useState, useCallback, type ChangeEvent } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Printer } from 'lucide-react';
 import toast from 'react-hot-toast';
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '../shared/ui/tabs';
-import { Card, CardContent } from '../shared/ui/card';
-import { Button } from '../shared/ui/button';
-import { Input } from '../shared/ui/input';
-import { Badge } from '../shared/ui/badge';
-import { Checkbox } from '../shared/ui/checkbox';
-import BarcodeScanner from '../shared/components/BarcodeScanner';
-import BarcodeGenerator from '../features/inventory/components/BarcodeGenerator';
-import { formatCurrency } from '../shared/lib/utils';
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '../../../shared/ui/tabs';
+import { Card, CardContent } from '../../../shared/ui/card';
+import { Button } from '../../../shared/ui/button';
+import { Input } from '../../../shared/ui/input';
+import { Badge } from '../../../shared/ui/badge';
+import { Checkbox } from '../../../shared/ui/checkbox';
+import BarcodeScanner from '../../../shared/components/BarcodeScanner';
+import BarcodeGenerator from '../../inventory/components/BarcodeGenerator';
+import { formatCurrency } from '../../../shared/lib/utils';
 import { useCartStore } from '../store/cartStore';
-import { useApiQuery } from '../shared/lib/apiQuery';
-import { useTransport } from '../shared/lib/transport/index';
-import { useTranslation } from '../shared/i18n/index';
+import { useApiQuery } from '../../../shared/lib/apiQuery';
+import { useTransport } from '../../../shared/lib/transport/index';
+import { useTranslation } from '../../../shared/i18n/index';
 import type { Product } from '@/types';
 
 export default function BarcodeTools() {
