+++ client/src/features/pos/pages/Register.test.tsx
@@ -2,9 +2,9 @@ import { describe, it, expect, beforeEach } from 'vitest';
 import type { ReactNode } from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { TransportProvider } from '../shared/lib/transport/index';
-import { createMemoryTransport, type MemoryTransport } from '../shared/lib/transport/memory';
-import { useSettingsStore } from '../shared/store/settingsStore';
+import { TransportProvider } from '../../../shared/lib/transport/index';
+import { createMemoryTransport, type MemoryTransport } from '../../../shared/lib/transport/memory';
+import { useSettingsStore } from '../../../shared/store/settingsStore';
 import type { RegisterReportData, RegisterSession } from '@/types';
 import RegisterPage from './Register';
